// web/src/api/generated/model/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Campus Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Announcements } from './announcements';
import type { DashboardNextClass } from './dashboardNextClass';
import type { DueItem } from './dueItem';
import type { LoanItem } from './loanItem';
import type { NoteItem } from './noteItem';
import type { ResultsStatus } from './resultsStatus';
import type { TodayClass } from './todayClass';

export interface Dashboard {
  today: TodayClass[];
  next_class: DashboardNextClass | null;
  due: DueItem[];
  announcements: Announcements;
  notes: NoteItem[];
  loans: LoanItem[];
  results: ResultsStatus;
}